# lib/api-spec/openapi.yaml
openapi: "3.0.0"
info:
  title: ACE-2026 API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.ace-2026.com
    description: Production Server
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - data
        - error
        - code
        - timestamp
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/keys:
    get:
      summary: Get status of all API keys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: API key statuses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      summary: Save or update an API key
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keyName: { type: string }
                keyValue: { type: string }
              required:
                - keyName
                - keyValue
      responses:
        "200":
          description: API key saved/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/keys/test:
    post:
      summary: Test connection for an API key
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keyName: { type: string }
                keyValue: { type: string }
              required:
                - keyName
                - keyValue
      responses:
        "200":
          description: Connection test successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Connection test failed
...